Aktueller Stand

+++ src/main/java/com/ibsys/scs/dto/neu/ArticleDto.java
@@ -5,7 +5,7 @@ import com.ibsys.scs.entities.neu.Article;
 
 public record ArticleDto(
         String id,
-        Double amount,
+        Integer amount,
         String number,
         String usedFor,
         String description,
